# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Fulcrum Core API
  description: A comprehensive cloud infrastructure management system designed to orchestrate and monitor distributed cloud resources across multiple providers.
  version: 1.0.0
  contact:
    name: Fulcrum Project
    url: https://fulcrumproject.org
  license:
    name: Apache License 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
...
servers:
  - url: https://api.fulcrum.testudosrl.dev/api/v1
    description: Development server

# Default security for all operations
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: oauth2
      description: Role-based authentication using bearer tokens
      flows:
        implicit:
          authorizationUrl: https://api.fulcrum.testudosrl.dev/oauth/authorize
          scopes:
            admin: System administrator with unrestricted access
            participant: Participant administrator that can act as both provider and consumer
            agent: Agent role with access to specific operations
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          description: HTTP status code
          example: 400
        error:
          type: string
          description: Error message
          example: "Invalid request parameters"
        details:
          type: string
          description: Additional error details
          example: "The field 'name' is required"
    properties.UUID:
      type: string
      format: uuid
      example: "123e4567-e89b-12d3-a456-426614174000"
    JSONObject:
      type: object
      additionalProperties: true
      description: A generic properties.JSON object

    # Participant schemas
...
    ParticipantResponse:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/properties.UUID"
        name:
          type: string
          example: "Test Participant"
        status:
          $ref: "#/components/schemas/ParticipantStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ParticipantStatus:
      type: string
      enum: [Enabled, Disabled]

    # Token schemas
...
    AgentStatus:
      type: string
      enum: [New, Connected, Disconnected, Error, Disabled]
...
    AgentResponse:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/properties.UUID"
        name:
          type: string
          example: "aws-agent-01"
        status:
          $ref: "#/components/schemas/AgentStatus"
        providerId:
          $ref: "#/components/schemas/properties.UUID"
        agentTypeId:
          $ref: "#/components/schemas/properties.UUID"
        tags:
          type: array
          items:
            type: string
          example: ["gpu", "high-memory", "ssd"]
        participant:
          $ref: "#/components/schemas/ParticipantResponse"
        agentType:
          $ref: "#/components/schemas/AgentTypeResponse"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    AgentTypeResponse:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/properties.UUID"
        name:
          type: string
          example: "Participant VM Agent"
        serviceTypes:
          type: array
          items:
            $ref: "#/components/schemas/ServiceTypeResponse"
          description: "Service types that this agent type can provide"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # Service schemas
    ServiceStatus:
      type: string
      enum:
        [
          Creating,
          Created,
          Starting,
          Started,
          Stopping,
          Stopped,
          HotUpdating,
          ColdUpdating,
          Deleting,
          Deleted,
        ]
    ServiceRequest:
      type: object
      required:
        - name
        - serviceTypeId
        - groupId
      properties:
        name:
          type: string
          example: "web-server-01"
        properties:
          $ref: "#/components/schemas/JSONObject"
        agentTags:
          type: array
          items:
            type: string
          example: ["gpu", "high-memory"]
          description: "Tags used for agent discovery when agentId is not specified"
        agentId:
          $ref: "#/components/schemas/properties.UUID"
          description: "Specific agent ID (optional - if not provided, agent discovery will use agentTags)"
        serviceTypeId:
          $ref: "#/components/schemas/properties.UUID"
        groupId:
          $ref: "#/components/schemas/properties.UUID"
    ServiceResponse:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/properties.UUID"
        name:
          type: string
          example: "web-server-01"
        currentStatus:
          $ref: "#/components/schemas/ServiceStatus"
        targetStatus:
          $ref: "#/components/schemas/ServiceStatus"
        currentProperties:
          $ref: "#/components/schemas/JSONObject"
        targetProperties:
          $ref: "#/components/schemas/JSONObject"
        failedAction:
          $ref: "#/components/schemas/ServiceAction"
        errorMessage:
          type: string
          example: "Failed to start service: insufficient resources"
        retryCount:
          type: integer
          example: 0
        resources:
          $ref: "#/components/schemas/JSONObject"
        externalId:
          type: string
        agentId:
          $ref: "#/components/schemas/properties.UUID"
        serviceTypeId:
          $ref: "#/components/schemas/properties.UUID"
        groupId:
          $ref: "#/components/schemas/properties.UUID"
        providerId:
          $ref: "#/components/schemas/properties.UUID"
        consumerId:
          $ref: "#/components/schemas/properties.UUID"
        agent:
          $ref: "#/components/schemas/AgentResponse"
        serviceType:
          $ref: "#/components/schemas/ServiceTypeResponse"
        group:
          $ref: "#/components/schemas/ServiceGroupResponse"
        consumerParticipant:
          $ref: "#/components/schemas/ParticipantResponse"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # Service Type schemas
...
    ServiceTypeResponse:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/properties.UUID"
        name:
          type: string
          example: "VM"
        propertySchema:
          $ref: "#/components/schemas/PropertySchema"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # Service Group schemas
    ServiceGroupRequest:
      type: object
      required:
        - name
        - consumerId
      properties:
        name:
          type: string
          example: "Web Servers"
        consumerId:
          $ref: "#/components/schemas/properties.UUID"
          description: "ID of the consumer participant that owns this service group"
    ServiceGroupResponse:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/properties.UUID"
        name:
          type: string
          example: "Web Servers"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # Action types
    ServiceAction:
      type: string
      enum:
        [
          ServiceCreate,
          ServiceStart,
          ServiceStop,
          ServiceHotUpdate,
          ServiceColdUpdate,
          ServiceDelete,
        ]

    # Job schemas
    # Using the same ServiceAction enum for job actions
...
    PageResponse:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
        totalItems:
          type: integer
        totalPages:
          type: integer
        pageSize:
          type: integer
        page:
          type: integer
    PropertySchema:
      type: object
      description: properties.JSON schema defining validation rules for service properties
      additionalProperties:
        $ref: "#/components/schemas/PropertyDefinition"
    PropertyDefinition:
      type: object
      properties:
        type:
          type: string
          enum: [string, integer, number, boolean, object, array]
          description: The data type of the property
        label:
          type: string
          description: Human-readable label for the property
        required:
          type: boolean
          description: Whether the property is required
          default: false
        default:
          description: Default value for the property
        validators:
          type: array
          items:
            $ref: "#/components/schemas/ValidatorDefinition"
          description: Array of validation rules
        properties:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/PropertyDefinition"
          description: Nested properties for object types
        items:
          $ref: "#/components/schemas/PropertyDefinition"
          description: Schema for array items
    ValidatorDefinition:
      type: object
      required:
        - type
        - value
      properties:
        type:
          type: string
          enum:
            [
              minLength,
              maxLength,
              pattern,
              enum,
              min,
              max,
              minItems,
              maxItems,
              uniqueItems,
            ]
          description: Type of validator
        value:
          description: Value for the validator (type depends on validator type)
...
    ValidatePropertiesResponse:
      type: object
      properties:
        valid:
          type: boolean
          description: Whether the properties are valid
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ValidationError"
          description: Array of validation errors (empty if valid)
    ValidationError:
      type: object
      properties:
        path:
          type: string
          description: Path to the property that failed validation
        message:
          type: string
          description: Error message describing the validation failure
...
paths:
...
  /services:
    get:
      summary: List services
      tags:
        - Services
      description: Retrieves a paginated list of services
      x-auth-permissions:
        - role: admin
          permission: all services
        - role: participant
          permission: services associated with its participant (as provider or consumer)
        - role: agent
          permission: services assigned to the agent
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: A paginated list of services
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PageResponse"
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: "#/components/schemas/ServiceResponse"
    post:
      summary: Create a service
      tags:
        - Services
      description: Creates a new service
      x-auth-permissions:
        - role: admin
          permission: always
        - role: participant
          permission: when acting as consumer
        - role: agent
          permission: not authorized
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceRequest"
      responses:
        "201":
          description: Service created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceResponse"
        "400":
          description: Invalid request or validation failed
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ErrorResponse"
                  - $ref: "#/components/schemas/ValidatePropertiesResponse"
  /services/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/properties.UUID"
    get:
      summary: Get a service
      tags:
        - Services
      description: Retrieves a specific service by ID
      x-auth-permissions:
        - role: admin
          permission: all services
        - role: participant
          permission: services associated with its participant (as provider or consumer)
        - role: agent
          permission: services assigned to the agent
      responses:
        "200":
          description: The service details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceResponse"
        "404":
          description: Service not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      summary: Update a service
      tags:
        - Services
      description: Updates an existing service
      x-auth-permissions:
        - role: admin
          permission: always
        - role: participant
          permission: services where it is the consumer participant
        - role: agent
          permission: not authorized
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "web-server-01"
                status:
                  $ref: "#/components/schemas/ServiceStatus"
                properties:
                  $ref: "#/components/schemas/JSONObject"
                resources:
                  $ref: "#/components/schemas/JSONObject"
                agentId:
                  $ref: "#/components/schemas/properties.UUID"
                serviceTypeId:
                  $ref: "#/components/schemas/properties.UUID"
                groupId:
                  $ref: "#/components/schemas/properties.UUID"
      responses:
        "200":
          description: Service updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceResponse"
        "400":
          description: Invalid request or validation failed
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ErrorResponse"
                  - $ref: "#/components/schemas/ValidatePropertiesResponse"
        "404":
          description: Service not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Delete a service
      tags:
        - Services
      description: Deletes a service by ID
      x-auth-permissions:
        - role: admin
          permission: always
        - role: participant
          permission: services where it is the consumer participant
        - role: agent
          permission: not authorized
      responses:
        "204":
          description: Service deleted successfully
        "404":
          description: Service not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /service-groups:
    get:
      summary: List service groups
      tags:
        - Services
      description: Retrieves a paginated list of service groups
      x-auth-permissions:
        - role: admin
          permission: all service groups
        - role: participant
          permission: service groups belonging to its participant
        - role: agent
          permission: not authorized
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: A paginated list of service groups
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PageResponse"
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: "#/components/schemas/ServiceGroupResponse"
    post:
      summary: Create a service group
      tags:
        - Services
      description: Creates a new service group
      x-auth-permissions:
        - role: admin
          permission: always
        - role: participant
          permission: for its participant
        - role: agent
          permission: not authorized
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceGroupRequest"
      responses:
        "201":
          description: Service group created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceGroupResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /service-groups/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/properties.UUID"
    get:
      summary: Get a service group
      tags:
        - Services
      description: Retrieves a specific service group by ID
      x-auth-permissions:
        - role: admin
          permission: all service groups
        - role: participant
          permission: service groups belonging to its participant
        - role: agent
          permission: not authorized
      responses:
        "200":
          description: The service group details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceGroupResponse"
        "404":
          description: Service group not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      summary: Update a service group
      tags:
        - Services
      description: Updates an existing service group
      x-auth-permissions:
        - role: admin
          permission: always
        - role: participant
          permission: service groups belonging to its participant
        - role: agent
          permission: not authorized
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceGroupRequest"
      responses:
        "200":
          description: Service group updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceGroupResponse"
        "404":
          description: Service group not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Delete a service group
      tags:
        - Services
      description: Deletes a service group by ID
      x-auth-permissions:
        - role: admin
          permission: always
        - role: participant
          permission: service groups belonging to its participant
        - role: agent
          permission: not authorized
      responses:
        "204":
          description: Service group deleted successfully
        "404":
          description: Service group not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Job endpoints
  #
...